+++ app/src/main/java/org/sopt/dosopttemplate/presentation/auth/SignUpActivity.kt
@@ -8,10 +8,10 @@ import org.sopt.dosopttemplate.R
 import org.sopt.dosopttemplate.data.entity.sign.SignInfo
 import org.sopt.dosopttemplate.databinding.ActivitySignupBinding
 import org.sopt.dosopttemplate.presentation.auth.LoginActivity.Companion.SIGN_INFO
-import sopt.uni.util.binding.BindingActivity
-import sopt.uni.util.extension.hideKeyboard
-import sopt.uni.util.extension.setOnSingleClickListener
-import sopt.uni.util.extension.showSnackbar
+import org.sopt.dosopttemplate.util.binding.BindingActivity
+import org.sopt.dosopttemplate.util.extension.hideKeyboard
+import org.sopt.dosopttemplate.util.extension.setOnSingleClickListener
+import org.sopt.dosopttemplate.util.extension.showSnackbar
 
 @AndroidEntryPoint
 class SignUpActivity : BindingActivity<ActivitySignupBinding>(R.layout.activity_signup) {
