+++ app/src/main/java/org/sopt/dosopttemplate/presentation/home/HomeDatePickerDialogFragment.kt
@@ -4,8 +4,8 @@ import android.os.Bundle
 import android.view.View
 import org.sopt.dosopttemplate.R
 import org.sopt.dosopttemplate.databinding.DatepickerDialogBinding
-import sopt.uni.presentation.BindingDialogFragment
-import sopt.uni.util.extension.setOnSingleClickListener
+import org.sopt.dosopttemplate.util.binding.BindingDialogFragment
+import org.sopt.dosopttemplate.util.extension.setOnSingleClickListener
 import java.util.Calendar
 
 class HomeDatePickerDialogFragment :
