+++ app/src/main/java/org/sopt/dosopttemplate/presentation/home/HomeVpAdapter.kt
@@ -1,28 +1,23 @@
 package org.sopt.dosopttemplate.presentation.home
 
 import android.view.LayoutInflater
-import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import coil.load
-import org.sopt.dosopttemplate.R
 import org.sopt.dosopttemplate.data.entity.home.Profile
 import org.sopt.dosopttemplate.databinding.ItemVpHomeFriendProfileBinding
 import org.sopt.dosopttemplate.databinding.ItemVpHomeMyprofileBinding
-import sopt.uni.util.extension.ItemDiffCallback
-import sopt.uni.util.extension.setOnSingleClickListener
+import org.sopt.dosopttemplate.util.extension.ItemDiffCallback
 
 class HomeVpAdapter(
     private val onClick: (Int) -> Unit,
     private val deleteUser: (String) -> Unit,
-) : ListAdapter<Profile, RecyclerView.ViewHolder>(
+) : ListAdapter<Profile, HomeVpViewHolder>(
     ItemDiffCallback<Profile>(
         onItemsTheSame = { old, new -> old.name == new.name },
         onContentsTheSame = { old, new -> old == new },
     ),
 ) {
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HomeVpViewHolder {
         return when (viewType) {
             MY_PROFILE_TYPE -> {
                 val binding = ItemVpHomeMyprofileBinding.inflate(
@@ -30,7 +25,7 @@ class HomeVpAdapter(
                     parent,
                     false,
                 )
-                MyProfileViewHolder(binding)
+                HomeVpViewHolder.MyProfileViewHolder(binding)
             }
 
             FRIEND_PROFILE_TYPE -> {
@@ -41,17 +36,17 @@ class HomeVpAdapter(
                 )
                 deleteUser
                 onClick
-                FriendProfileViewHolder(binding)
+                HomeVpViewHolder.FriendProfileViewHolder(binding, onClick, deleteUser)
             }
 
             else -> throw IllegalArgumentException(UNKNOWN_TYPE)
         }
     }
 
-    override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: HomeVpViewHolder, position: Int) {
         when (holder) {
-            is MyProfileViewHolder -> holder.onBind(getItem(position) as Profile.MyProfile)
-            is FriendProfileViewHolder -> holder.onBind(getItem(position) as Profile.FriendProfile)
+            is HomeVpViewHolder.MyProfileViewHolder -> holder.onBind(getItem(position) as Profile.MyProfile)
+            is HomeVpViewHolder.FriendProfileViewHolder -> holder.onBind(getItem(position) as Profile.FriendProfile)
         }
     }
 
@@ -60,44 +55,6 @@ class HomeVpAdapter(
             is Profile.MyProfile -> MY_PROFILE_TYPE
             is Profile.FriendProfile -> FRIEND_PROFILE_TYPE
             else -> MY_PROFILE_TYPE
-        }
-    }
-
-    inner class MyProfileViewHolder(private val binding: ItemVpHomeMyprofileBinding) :
-        RecyclerView.ViewHolder(binding.root) {
-        fun onBind(myProile: Profile.MyProfile) {
-            binding.ivVpHomeMyprofile.load(myProile.profileImage)
-            binding.tvVpHomeMyName.text = myProile.name
-        }
-    }
-
-    inner class FriendProfileViewHolder(private val binding: ItemVpHomeFriendProfileBinding) :
-        RecyclerView.ViewHolder(binding.root) {
-        fun onBind(friendProfile: Profile.FriendProfile) {
-            with(binding) {
-                ivVpHomeFriendProfile.load(friendProfile.profileImage)
-                tvVpHomeFriendName.text = friendProfile.name
-
-                clVpHomeFriend.setOnSingleClickListener {
-                    onClick(absoluteAdapterPosition)
-                }
-
-                clVpHomeFriend.setOnLongClickListener {
-                    deleteUser(friendProfile.name)
-                    true
-                }
-
-                if (friendProfile.isTodayBirthday) {
-                    ivVpHomeFriendBirthdayCake.visibility = View.VISIBLE
-                    tvVpHomeFriendGift.visibility = View.VISIBLE
-                    tvVpHomeFriendGift.text = binding.root.context.getString(R.string.home_gift)
-                }
-
-                if (friendProfile.isMusicRegist) {
-                    clVpHomeFriendMusic.visibility = View.VISIBLE
-                    tvVpHomeFriendMusicName.text = friendProfile.music
-                }
-            }
         }
     }
